requirements.txt
python-dotenv
llama-index 
llama-index-embeddings-nebius 
llama-index-vector-stores-chroma 
sqlalchemy
llama-index-llms-nebius
openai
tqdm